[ADDITION/FIX]
Added the id to ClienCredentials, this is so that the registered clients hashmap has the correct usernames and passwords. Changed the layout of the client app, so that it can login or register (for the first time). Data storage of the client login is stored correctly. Added an enum to find a difference between logging in and registering. Changed a couple of class modifiers to public so that everything could be connected in a correct fashion.